Added some experimental rules for long-distance pronoun resolution (search for -1W):
Looking across sentence border in order to find feminine names or
nouns (having "han" as fallback (?)).
Note. W added only to *-1, not to *1.
Also added a rule referring to last sentence subject.

The tag (ant f) may be wrong (sem_fem), it was part of debugging:
we are now able to translate "son" to "hun" with remote subject,
but if the subordinate subject is empty
Márjá lea dáppe. Son lea ilus, ja oaidná dál maid man hávski dáppe lea.

Here, "Son" becomes "Hun", as required, but the inserted subject
of "oaidná" is Han/hun, as before, since this now is done in
transfer (we might actually do an APPEND rule here in .lex as well).